modify
git-svn-id: https://svn.baidu.com/myspace/sep/chenghuige/melt-train@124438 c4ccc3fd-b13d-49bf-a705-f618d8f4424a

+++ include/Trainers/FastRank/FastRank.h
@@ -186,7 +186,7 @@ namespace gezi {
 			{
 				return InitValidScores;
 			}
-			for (int i = 0; (!TestSets.empty()) && (i < TestSets.size()); i++)
+			for (size_t i = 0; (!TestSets.empty()) && (i < TestSets.size()); i++)
 			{
 				if (&set == &TestSets[i])
 				{
